Dispatch event on commit failure with original error
Wehn commit fails, The entitymanager gets closed. Any later action (persist, flush) will throw an entityManager closed exception.
But the original message is lost.
We dispatch this event so that we can listen to it in our applications and log the original reason why the entity manager is getting closed.